vault/requirements.txt
atlassian-python-api==3.41.11
azure-common==1.1.28
azure-core==1.30.1
azure-search-documents==11.4.0b8
bcrypt==4.2.1
docx2python==3.3.0
fastapi-mail==1.4.2
gradio==5.29.1
langchain==0.1.16
langchain-community==0.0.38
msal==1.31.0
mysql-connector-python==9.1.0
openai==1.79.0
pandas==2.2.2
psycopg2-binary==2.9.9
PyPDF2==3.0.1
python-docx==1.1.2
python-ldap==3.4.4
regex==2024.5.15
requests==2.31.0
tqdm==4.66.2
uvicorn==0.34.2
fastapi==0.115.12
python-dotenv==1.0.1
qdrant-client==1.14.2
scikit-learn==1.6.1
scipy==1.15.2
sentence-transformers==4.1.0
sqlalchemy==2.0.41
supabase==2.15.1
websockets==14.2